Added delete delivery price when delete product

+++ app/Http/Controllers/MediaController.php
@@ -13,7 +13,24 @@ use Intervention\Image\Facades\Image;
 
 class MediaController extends Controller
 {
-    public function UploadMedia(Request $request)
+
+    public function uploadMedia(Request $request){
+        if(config('app.PHOTO_DRIVER') == 'local'){
+            return $this->uploadLocal($request);
+        }elseif (config('app.PHOTO_DRIVER') == 's3'){
+            return $this->uploadToS3($request);
+        }
+    }
+
+    public function destroyMedia(Request $request){
+        if(config('app.PHOTO_DRIVER') == 'local'){
+            return $this->destroyLocal($request);
+        }elseif (config('app.PHOTO_DRIVER') == 's3'){
+            return $this->destroyS3($request);
+        }
+    }
+
+    private function uploadLocal(Request $request)
     {
         // Validate input
         $validator = Validator::make($request->all(), [
@@ -75,7 +92,7 @@ class MediaController extends Controller
         }
     }
 
-    public function uploadToS3(Request $request){
+    private function uploadToS3(Request $request){
         // Validate input
         $validator = Validator::make($request->all(), [
             'uploaded_file' => 'required|image|max:400'
@@ -89,7 +106,7 @@ class MediaController extends Controller
             $image = $request->file('uploaded_file');
             $url = 'https://s3.' . env('AWS_DEFAULT_REGION') . '.amazonaws.com/' . env('AWS_BUCKET') . '/';
             $path = 'images/';
-            $thumb_path = 'images/thumb/';
+            $thumb_path = 'images/thumbnail/';
 
             $file_ext = $request->file('uploaded_file')->clientExtension();
             $file_name = md5(uniqid());
@@ -155,28 +172,54 @@ class MediaController extends Controller
          }
     }
 
-
-
-
-
-    public function destroy(Request $request)
+    private function destroyLocal(Request $request)
     {
-        $media_item = Media::where('id', $request->mediaId)->first();
+        if(!Media::where('id', $request->mediaId)->exists()){
+            return response()->json(['status' => 400, 'message' => 'Failed to delete Media. File does not exist'], 400);
+        }
+
+        $media_item = Media::where('id', $request->mediaId);
+        $file = $media_item->first()->file;
+
         try {
-            $path_file = public_path('/' . $media_item->file);
-            $thumb_path_file = public_path('thumbnail/' . $media_item->file);
-            $file_exists = File::exists($media_item->file);
-            $thumb_file_exists = File::exists('thumbnail/' . $media_item->file);
-            if ($file_exists) {
+            $path_file = public_path('/' . $file);
+            $thumb_path_file = public_path('thumbnail/' . $file);
+
+            if (File::exists($file)) {
                 File::delete($path_file);
             }
-            if ($thumb_file_exists) {
+
+            if (File::exists('thumbnail/' . $file)) {
                 File::delete($thumb_path_file);
             }
             $media_item->delete();
             return response()->json(['status' => 200, 'message' => 'Media deleted'], 200);
         } catch (\Exception $e) {
             return response()->json(['status' => 400, 'message' => 'Failed to delete Media'], 400);
+        }
+    }
+
+    private function destroyS3(Request $request)
+    {
+        if(!Media::where('id', $request->mediaId)->exists()){
+            return response()->json(['status' => 400, 'message' => 'Failed to delete Media. File does not exist'], 400);
+        }
+
+        $media_item = Media::where('id', $request->mediaId);
+        $file = $media_item->first()->file;
+        try {
+            if (Storage::disk('s3')->exists('images/'.$file)) {
+                //DELETE PHOTO
+                Storage::disk('s3')->delete('images/'.$file);
+            }
+            if (Storage::disk('s3')->exists('images/thumbnail/'.$file)) {
+                //DELETE THUMB
+                Storage::disk('s3')->delete('images/thumbnail/'.$file);
+            }
+            $media_item->delete();
+            return response()->json(['status' => 200, 'message' => 'Media deleted'], 200);
+        } catch (\Exception $e) {
+            return response()->json(['status' => 400, 'message' => 'Failed to delete Media '.$e->getMessage()], 400);
         }
     }
